backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
Werkzeug==3.0.1
requests==2.31.0
scipy==1.11.4

